requirements.txt
Pillow
django>=4.2
python-dotenv
psycopg2-binary
djangorestframework
markdown
django-filter